Cakev13 big changes

+++ src/main/java/edu/bhscs/PTSA.java
@@ -2,7 +2,7 @@ package edu.bhscs;
 
 public class PTSA {
   Clerk myClerk = new Clerk();
-  Integer ptsaMoney = myClerk.getClerkCash();
+  Double ptsaMoney = myClerk.getClerkCash();
 
   public String donate(){
     return("The clerk donated " + ptsaMoney + " money to the PTSA");
